requirements.txt
asgiref==3.7.2
beautifulsoup4==4.11.2
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
Django==4.2.5
django-filter==23.3
django-jsonfield==1.4.1
django-model-utils==4.3.1
django-modelcluster==6.0
django-notifications-hq==1.8.2
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
heroicons==2.5.0
html5lib==1.1
idna==3.4
jsonfield==3.1.0
l18n==2021.3
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pillow-heif==0.13.0
plotly==5.17.0
python-dateutil>=2.6.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
swapper==1.3.0
telepath==0.3.1
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wagtail==5.1.1
webencodings==0.5.1
Willow==1.6.2
whitenoise